requirements.txt
asyncpg==0.29.0
bcrypt==4.2.0
fastapi==0.115.0
pydantic==2.9.2
SQLAlchemy==2.0.35
uvicorn==0.30.6
python-dotenv==1.0.1